ml-server/requirements.txt
fastapi
uvicorn
pillow
numpy
torch
torchvision
transformers
python-multipart

joblib==1.4.2
scikit-learn==1.5.2
pydantic==2.8.2